+++ completion.go
@@ -105,8 +105,7 @@ func (p *Parser) WriteNamedCompletion(w io.Writer, shell CompletionShell, comman
 	args=("${COMP_WORDS[@]:1:$COMP_CWORD}")
 	cur="${COMP_WORDS[COMP_CWORD]}"
 
-	local IFS=$'\n'
-	COMPREPLY=($(GO_FLAGS_COMPLETION=1 ${COMP_WORDS[0]} "${args[@]}"))
+	mapfile -t COMPREPLY < <(GO_FLAGS_COMPLETION=1 ${COMP_WORDS[0]} "${args[@]}")
 
 	if [[ "$cur" == --*=* || "$cur" == -*=* ]]; then
 		compopt -o nospace 2>/dev/null
@@ -153,15 +152,15 @@ Register-ArgumentCompleter -Native -CommandName $__goFlagsCommand -ScriptBlock {
 	}
 
 	$exe = $elements[0]
-	$args = @()
+	$__goFlagsArgs = @()
 	if ($elements.Count -gt 1) {
-		$args = $elements[1..($elements.Count - 1)]
+		$__goFlagsArgs = $elements[1..($elements.Count - 1)]
 	}
 
 	$prev = $env:GO_FLAGS_COMPLETION
 	$env:GO_FLAGS_COMPLETION = '1'
 	try {
-		$items = & $exe @args
+		$items = & $exe @__goFlagsArgs
 	} finally {
 		if ($null -ne $prev) {
 			$env:GO_FLAGS_COMPLETION = $prev